% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{appliquer_casse}
\alias{appliquer_casse}
\title{Appliquer la casse d'une premiere chaine de caracteres vers une seconde chaine de caracteres}
\usage{
appliquer_casse(libelle, libelle_casse)
}
\arguments{
\item{libelle}{Un vecteur de type caractère.}

\item{libelle_casse}{Un vecteur de type caractère servant de modèle pour la casse.}
}
\value{
Un vecteur de type caractère dont la casse est mise à jour.
}
\description{
Appliquer la casse d'une première chaine de caractères vers une seconde chaine de caractères.
}
\examples{
# Un exemple
caractr::appliquer_casse(c("test1", "test2"), libelle_casse = c("aaAaa", "bBBbb"))

}

