% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{camel_to_snake_case}
\alias{camel_to_snake_case}
\title{Passage d'un nom de champ de la casse camel a snake}
\usage{
camel_to_snake_case(char)
}
\arguments{
\item{char}{Un vecteur de chaines de caractères à la casse camel.}
}
\value{
Un vecteur de chaines de caractères à la casse snake.
}
\description{
Passage d'un nom de champ de la casse camel à snake.
}
\examples{
caractr::camel_to_snake_case(c("emploiNN1Type", "rechercheEmploi"))

}
