% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{conv_ods_na_vide}
\alias{conv_ods_na_vide}
\title{Open data - Convertir les NA d'une chaine de caracteres vers une chaine vide}
\usage{
conv_ods_na_vide(libelle)
}
\arguments{
\item{libelle}{Un vecteur de type caractère.}
}
\value{
Un vecteur de type caractère dont les \code{NA} sont transformés en \code{""}.
}
\description{
Convertir les \code{NA} d'une chaine de caractères vers une chaine vide.\cr
}
\details{
Utilisé conjointement avec la fonction \code{conv_pv_vide()} pour créer les champs multivalués vides des fichiers en open data.
}
\examples{
caractr::conv_ods_na_vide(c("Université de Franche-Comté", NA_character_))

}
