% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{conv_ods_pv_na}
\alias{conv_ods_pv_na}
\title{Open data - Convertir les points-virgules seuls vers NA}
\usage{
conv_ods_pv_na(libelle)
}
\arguments{
\item{libelle}{Un vecteur de type caractère.}
}
\value{
Un vecteur de type caractère dont les points-virgule seuls sont transformés en \code{NA}.
}
\description{
Convertir les points-virgules seuls vers \code{NA}.\cr
}
\details{
Utilisé conjointement avec la fonction \code{conv_na_vide()} pour créer les champs multivalués vides des fichiers en open data.
}
\examples{
conv_ods_pv_na(c(";;;", ";CNRS;;"))

}

