% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{maj_accent}
\alias{maj_accent}
\title{Mise a jour de mots non-accentues}
\usage{
maj_accent(libelle)
}
\arguments{
\item{libelle}{Un vecteur de type caractère.}
}
\value{
Un vecteur de type caractère dont les mots sont accentués.

Jeu de données source : \code{caractr::mot_accent}.\cr
Il est créé à partir d'une table source SAS (projet "Text mining").\cr
}
\description{
Mise à jour de mots non-accentués.
}
\examples{
# Un exemple qui fonctionne bien
caractr::maj_accent("Ecole superieure de commerce de Troyes")

# Un exemple qui fonctionne à moitié
caractr::maj_accent("Universite de Franche-Comte")

}
