% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mots_vides.R
\name{maj_casse}
\alias{maj_casse}
\title{Mise a jour de la casse en prenant en compte les mots vides}
\usage{
maj_casse(libelle, excepte = NULL, code_langue = "fr")
}
\arguments{
\item{libelle}{Un vecteur de type caractère.}

\item{excepte}{Vecteur de mots à ne pas mettre à jour.}

\item{code_langue}{Code de la langue des mots vides.}
}
\value{
Un vecteur de type caractère dont tous les mots commencent par une majuscule sauf les mots vides de la langue sélectionnée.
}
\description{
Mise en majuscule de la première lettre de chaque mot sauf les mots_vides.
}
\examples{
# un premier exemple
caractr::maj_casse(c("boulevard D'argonne", "1 rue descartes"))

# Un exemple avec un mot non-sélectionné
caractr::maj_casse(c("boulevard D'argonne", "1 rue descartes"), excepte = "rue")

# Un exemple en anglais
caractr::maj_casse(c("GENETICS AND IMMUNOLOGY OF PARASITIC DISEASES",
  "RESEARCH INSTITUTE  IN HEALTH"), code_langue = "en")

}

