% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordonnees.R
\name{nettoyer_numero_telephone}
\alias{nettoyer_numero_telephone}
\title{Nettoyer des numeros de telephone}
\usage{
nettoyer_numero_telephone(numero)
}
\arguments{
\item{numero}{Un vecteur de numéros.}
}
\value{
Un vecteur de numéros nettoyés.
}
\description{
Nettoyer des numéros de téléphone.\cr
}
\examples{
caractr::nettoyer_numero_telephone("06.00.00.00.00")

}
