% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{normaliser_char}
\alias{normaliser_char}
\title{Normaliser une chaine de caracteres (utilisation pour un nom de champ dans une table)}
\usage{
normaliser_char(libelle)
}
\arguments{
\item{libelle}{Un vecteur de type caractère.}
}
\value{
Un vecteur de type caractère contenant les libellés normalisés.
}
\description{
Normaliser une chaine de caractères (utilisation pour un nom de champ dans une table).
}
\examples{
caractr::normaliser_char(c("Type d'unité Sirus : entreprise profilée ou unité légale", "Nic du siège"))

}
