% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{paste2}
\alias{paste2}
\title{Derive de la fonction paste() avec un parametre de suppression des NA}
\usage{
paste2(..., sep = " ", collapse = NULL, na.rm = TRUE)
}
\arguments{
\item{...}{Des chaines de caractère à concaténer.}
}
\value{
Une chaine de caractère concaténée sans \code{NA}.
}
\description{
Dérivé de la fonction \code{paste()} avec un paramètre de suppression des \code{NA}.\cr
}
\details{
Trouvé sur : \url{http://stackoverflow.com/questions/13673894/suppress-nas-in-paste}
}
\examples{
# Avec la fonction paste() classique
paste("chaine1", NA_character_, "chaine2")

# Avec la fonction paste2()
caractr::paste2("chaine1", NA_character_, "chaine2")

}
