% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{paste_na}
\alias{paste_na}
\title{Derive de la fonction paste() mais retourne NA si un seul des elements est a NA}
\usage{
paste_na(..., sep = " ", collapse = NULL)
}
\arguments{
\item{...}{Des chaines de caractère à concaténer.}
}
\value{
Une chaine de caractère concaténée sans \code{NA}.
}
\description{
Dérivé de la fonction \code{paste()} mais retourne \code{NA} si un seul des elements est a \code{NA}\cr
}
\examples{
# Avec la fonction paste() classique
paste("chaine1", NA_character_, "chaine2")

# Avec la fonction paste_na()
caractr::paste_na("chaine1", NA_character_, "chaine2")

}

