% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mots_vides.R
\name{prx_mots_vides}
\alias{prx_mots_vides}
\title{Obtenir la liste des mots vides (fr ou en) sous forme d'expression reguliere}
\usage{
prx_mots_vides(langue = "fr", excepte = NULL, selection = NULL)
}
\arguments{
\item{langue}{Code de la langue des mots vides.}

\item{excepte}{Vecteur de mots-vides à ne pas sélectionner.}

\item{selection}{Vecteur de mots-vides à sélectionner.}
}
\value{
Une expression régulière sous forme de chaine de caractères.\cr

Jeu de données source : \code{caractr::data_mot_vide}.\cr
Il est créé à partir de la table "Mot_vide" de la base Access "Tables_ref.accdb" (projet "Text mining").
}
\description{
Obtenir la liste des mots vides (fr ou en) sous forme d'expression régulière.
}
\examples{
# Les mots vides français
caractr::prx_mots_vides()

# Les mots vides français sauf les mots "au" et "aux"
caractr::prx_mots_vides(excepte = c("au", "aux"))

# Les mots vides français "au" et "aux"
caractr::prx_mots_vides(selection = c("au", "aux"))

# Les mots vides anglais
caractr::prx_mots_vides(langue = "en")

}
