% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mots_vides.R
\name{prx_stopwords}
\alias{prx_stopwords}
\title{Get a list of stop words as a regular expression.}
\usage{
prx_stopwords(language = "fr", drop = NULL, keep = NULL)
}
\arguments{
\item{language}{Stop word language.}

\item{drop}{Stop words to drop.}

\item{keep}{Stop words to keep.}
}
\value{
A regular expression as string character.
}
\description{
Get a list of stop words as a regular expression.
}
\details{
Source dataset : \code{caractr::stopwords}.
}
\examples{
# French stopwords
caractr::prx_stopwords()

# French stopwords except "au" and "aux"
caractr::prx_stopwords(drop = c("au", "aux"))

# French stopwords but only "au" and "aux"
caractr::prx_stopwords(keep = c("au", "aux"))

# English stopwords
caractr::prx_stopwords(language = "en")

}
\keyword{internal}
