% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{sans_accent}
\alias{sans_accent}
\title{Supprimer les accents d'une chaine de caracteres}
\usage{
sans_accent(libelle)
}
\arguments{
\item{libelle}{Un vecteur de type caractère.}
}
\value{
Un vecteur de type caractère sans accent.
}
\description{
Supprimer les accents d'une chaine de caractères.\cr
}
\details{
Trouvé sur : \url{http://stackoverflow.com/questions/20495598/replace-accented-characters-in-r-with-non-accented-counterpart-utf-8-encoding}
}
\examples{
caractr::sans_accent("Université de Franche-Comté")

}

