% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{sans_ponctuation}
\alias{sans_ponctuation}
\title{Supprimer la ponctuation d'une chaine de caracteres}
\usage{
sans_ponctuation(libelle, remplacement = " ")
}
\arguments{
\item{libelle}{Un vecteur de type caractère.}

\item{remplacement}{Le caractère de remplacement de la ponctuation.}
}
\value{
Un vecteur de type caractère sans ponctuation.
}
\description{
Supprimer la ponctuation d'une chaine de caractères
}
\examples{
caractr::sans_ponctuation("Université Paris-Est Créteil Val-de-Marne")
caractr::sans_ponctuation("Université Paris-Est Créteil Val-de-Marne", remplacement = "_")

}
