% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{A list of stop words.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 145 rows and 2 columns.
}
\usage{
stopwords
}
\description{
\preformatted{## Rows: 145
## Columns: 2
## $ language <chr> "fr", "fr", "fr", "fr", "fr", "fr", "fr", "fr", "fr", "fr", "fr", "fr", "fr", ...
## $ stopword <chr> "au", "aux", "avec", "ce", "ces", "dans", "de", "des", "du", "elle", "en", "et...
}
}
\keyword{datasets}
