% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopword.R
\name{str_add_case}
\alias{str_add_case}
\title{Update a character vector case according to stopwords.}
\usage{
str_add_case(string, drop = NULL, language = "fr")
}
\arguments{
\item{string}{Input character vector.}

\item{drop}{Stop words to drop.}

\item{language}{Stop word language.}
}
\value{
A character vector.
}
\description{
Update a character vector case according to stopwords.
}
\examples{
caractr::str_add_case(c("boulevard D'argonne", "1 rue descartes"))

# An example with a dropped word
caractr::str_add_case(c("boulevard D'argonne", "1 rue descartes"), drop = "rue")

# An English example
caractr::str_add_case(c("GENETICS AND IMMUNOLOGY OF PARASITIC DISEASES",
  "RESEARCH INSTITUTE  IN HEALTH"), language = "en")

}
