% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{str_age_fr}
\alias{str_age_fr}
\title{Get age in a French litteral format.}
\usage{
str_age_fr(age)
}
\arguments{
\item{age}{A numeric vector of age.}
}
\value{
A character vector of age in French litteral format.
}
\description{
Get age in a French litteral format.
}
\examples{
caractr::str_age_fr(25.64048)

}
