% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{str_age_litteral}
\alias{str_age_litteral}
\title{Get age in a litteral format.}
\usage{
str_age_litteral(age)
}
\arguments{
\item{age}{A numeric vector of age.}
}
\value{
A character vector of age in litteral format.
}
\description{
Get age in a litteral format.
}
\examples{
caractr::str_age_litteral(25.64048)

}
