% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{str_apostrophe}
\alias{str_apostrophe}
\title{remplace l'apostrophe Windows par l'apostrophe ASCII}
\usage{
str_apostrophe(char)
}
\arguments{
\item{char}{Un vecteur de chaines de caractères.}
}
\value{
Un vecteur de chaines de caractères sans apostrophe Windows.
}
\description{
remplace l'apostrophe Windows par l'apostrophe ASCII.
}
\examples{
caractr::str_apostrophe("Un emploi qui n’a pas de lien avec vos études")

}
