% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_apply_case}
\alias{str_apply_case}
\title{Apply case from a target character to another string.}
\usage{
str_apply_case(string, target)
}
\arguments{
\item{string}{Input character vector.}

\item{target}{A case target character vector.}
}
\value{
A character vector.
}
\description{
Apply case from a target character to another string.
}
\examples{
caractr::str_apply_case(c("test1", "test2"), target = c("aaAaa", "bBBbb"))

}
