% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{str_camel_to_snake_case}
\alias{str_camel_to_snake_case}
\title{Conversion from a camel case character to a snake case character.}
\usage{
str_camel_to_snake_case(string)
}
\arguments{
\item{string}{Input character vector.}
}
\value{
A character vector.
}
\description{
Conversion from a camel case character to a snake case character.
}
\examples{
caractr::str_camel_to_snake_case(c("emploiNN1Type", "rechercheEmploi"))

}
