% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{str_capitalise}
\alias{str_capitalise}
\title{Capitalise the first letter in a string.}
\usage{
str_capitalise(string)
}
\arguments{
\item{string}{Input character vector.}
}
\value{
A character vector.
}
\description{
Capitalise the first letter in a string.
}
\examples{
caractr::str_capitalise("université de Caen Normandie")

}
