% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{str_capitaliser}
\alias{str_capitaliser}
\title{Mettre une capitale a la premiere lettre d'une chaine de caracteres}
\usage{
str_capitaliser(libelle)
}
\arguments{
\item{libelle}{Un vecteur de type caractère.}
}
\value{
Un vecteur de type caractère dont la première lettre de chaque élément est une majuscule.
}
\description{
Mettre une capitale à la première lettre d'une chaine de caractères.
}
\examples{
caractr::str_capitaliser("université de Caen Normandie")

}

