% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordonnees.R
\name{str_clean_phone_number}
\alias{str_clean_phone_number}
\title{Clean phone numbers.}
\usage{
str_clean_phone_number(phone_number)
}
\arguments{
\item{phone_number}{A vector of phone numbers.}
}
\value{
A vector of cleaned phone numbers.
}
\description{
Clean phone numbers.
}
\examples{
caractr::str_clean_phone_number("06.00.00.00.00")

}
