% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R
\name{str_conv_number_letter}
\alias{str_conv_number_letter}
\title{Convert a numeric vector to full letters.}
\usage{
str_conv_number_letter(x, type = "letter", female = FALSE, language = "fr")
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{Choice between : letter|ieme|ieme_number}

\item{female}{If \code{TRUE}, female form.}

\item{language}{French language only available.}
}
\description{
Convert a numeric vector to full letters.
}
\examples{
caractr::str_conv_number_letter(1:10)
caractr::str_conv_number_letter(1, female = TRUE)

}
