% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_empty_to_na}
\alias{str_empty_to_na}
\title{Replace Empty character with NA.}
\usage{
str_empty_to_na(string)
}
\arguments{
\item{string}{Input character vector.}
}
\value{
A character vector.
}
\description{
Replace Empty character with NA.
}
\examples{
caractr::str_empty_to_na(c("a", "", "c"))

}
