% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{str_fichier}
\alias{str_fichier}
\title{Normaliser des chaines de caractere en tant que nom de fichier}
\usage{
str_fichier(char)
}
\arguments{
\item{char}{Un vecteur de type caractère.}
}
\value{
Un vecteur de type caractère prêts à être utilisés en tant que nom de fichier.
}
\description{
Normaliser des chaines de caractère en tant que nom de fichier.
}
\examples{

}
