% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_line_break}
\alias{str_line_break}
\title{Cut a character string over multiple lines}
\usage{
str_line_break(string, nchar_max, collapse = "\\n")
}
\arguments{
\item{string}{Input character vector.}

\item{nchar_max}{The maximum number of character per line.}

\item{collapse}{A character string to separate the results.}
}
\value{
A character vector.
}
\description{
Cut a character string over multiple lines
}
\examples{
caractr::str_line_break(c("A very very very very very very long character string", "test"),
 nchar_max = 40)

}
