% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{str_normalise_field}
\alias{str_normalise_field}
\title{Normalise a string for a use as table field name.}
\usage{
str_normalise_field(string)
}
\arguments{
\item{string}{Input character vector.}
}
\value{
A character vector.
}
\description{
Normalise a string for a use as table field name.
}
\examples{
caractr::str_normalise_field(c("Type d'unité Sirus : entreprise profilée ou unité légale", "Nic du siège"))

}
