% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_normalise_file}
\alias{str_normalise_file}
\title{Normalise a string for a use as file name.}
\usage{
str_normalise_file(string, replace_slash = TRUE)
}
\arguments{
\item{string}{Input character vector.}

\item{replace_slash}{Replace also \code{/} character.}
}
\value{
A character vector.
}
\description{
Normalise a string for a use as file name.
}
