% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{str_number_fr}
\alias{str_number_fr}
\title{Conversion from a number to a French formatted number.}
\usage{
str_number_fr(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A French formatted number.
}
\description{
Includes comma as decimal mark and non-breaking space between thousands.
}
\examples{
caractr::str_number_fr(1854.2)

}
