% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{str_paste}
\alias{str_paste}
\title{Equivalent to base paste but with an extra na.rm parameter}
\usage{
str_paste(..., sep = " ", collapse = NULL, na.rm = TRUE)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}

\item{sep}{a character string to separate the terms.}

\item{collapse}{an optional character string to separate the results.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
A character vector of the concatenated values.
}
\description{
Found on : \url{http://stackoverflow.com/questions/13673894/suppress-nas-in-paste}
}
\examples{
paste("chaine1", NA_character_, "chaine2")
caractr::str_paste("chaine1", NA_character_, "chaine2")

}
