% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_percent}
\alias{str_percent}
\title{Conversion from a numeric percentage to a character label.}
\usage{
str_percent(x, digits = 1, symbol = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Integer indicating the number of decimal places.}

\item{symbol}{Display of "\%".}
}
\value{
A character vector.
}
\description{
Conversion from a numeric percentage to a character label.
}
\examples{
caractr::str_percent(0.1)

}
