% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{str_percent_fr}
\alias{str_percent_fr}
\title{Conversion from a numeric percentage to a French formatted label.}
\usage{
str_percent_fr(x, digits = 0, suffix = "<U+202F>\%", sign = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Integer indicating the number of decimal places.}

\item{suffix}{Suffix after the percent value.}

\item{sign}{Display of "+" and "-".}
}
\value{
A French formatted label.
}
\description{
Conversion from a numeric percentage to a French formatted label.
}
\examples{
caractr::str_percent_fr(0.1)

}
