% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_remove_accent}
\alias{str_remove_accent}
\title{Remove accents in a string.}
\usage{
str_remove_accent(string)
}
\arguments{
\item{string}{Input character vector.}
}
\value{
A character vector.
}
\description{
Found on : \url{http://stackoverflow.com/questions/20495598/replace-accented-characters-in-r-with-non-accented-counterpart-utf-8-encoding}
}
\examples{
caractr::str_remove_accent("Universit\\u00E9 de Franche-Comt\\u00E9")

}
