% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_remove_punct}
\alias{str_remove_punct}
\title{Remove punctuation in a string.}
\usage{
str_remove_punct(string, replacement = " ")
}
\arguments{
\item{string}{Input character vector.}

\item{replacement}{A character vector of replacements.}
}
\value{
A character vector.
}
\description{
Remove punctuation in a string.
}
\examples{
caractr::str_remove_punct("Universit\\u00E9 Paris-Est Cr\\u00E9teil Val-de-Marne")
caractr::str_remove_punct("Universit\\u00E9 Paris-Est Cr\\u00E9teil Val-de-Marne", replacement = "_")

}
