% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_replace_win_quote}
\alias{str_replace_win_quote}
\title{Replace Windows quote by the ASCII quote}
\usage{
str_replace_win_quote(string)
}
\arguments{
\item{string}{Input character vector.}
}
\value{
A character vector.
}
\description{
Replace Windows quote by the ASCII quote
}
\examples{
caractr::str_replace_win_quote("Un emploi qui n\uE28099a pas de lien avec vos \u00E9tudes")

}
