% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caractr.R
\name{str_saut_ligne}
\alias{str_saut_ligne}
\title{Decoupage d'une chaine de caracteres sur plusieurs lignes}
\usage{
str_saut_ligne(char, n_char_max, collapse = "\\n")
}
\arguments{
\item{char}{Un vecteur de chaines de caractères.}

\item{n_char_max}{Un nombre de caractères maximum par ligne.}

\item{collapse}{Le séparateur pour le saut de ligne.}
}
\value{
Un vecteur de chaines de caractères.
}
\description{
Découpage d'une chaine de caractères sur plusieurs lignes.
}
\examples{
caractr::str_saut_ligne(c("Une très très très très très très longue chaine de caractère", "test"), nchar = 40)

}
