% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_thousands}
\alias{str_thousands}
\title{Conversion from a numeric to a character with thousands separated by a non-breaking space.}
\usage{
str_thousands(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector.
}
\description{
Conversion from a numeric to a character with thousands separated by a non-breaking space.
}
\examples{
caractr::str_percent(1854)

}
