% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_today}
\alias{str_today}
\title{Get today date in several formats.}
\usage{
str_today(format = "file")
}
\arguments{
\item{format}{Date format to use : \code{file} or \code{litteral}.}
}
\value{
Date of the day.
}
\description{
Get today date in several formats.
}
\examples{
caractr::str_today()
caractr::str_today("litteral")

}
