% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{str_today_fr}
\alias{str_today_fr}
\title{Get today French date in several formats.}
\usage{
str_today_fr(format = "file")
}
\arguments{
\item{format}{Date format to use : \code{file} or \code{litteral}.}
}
\value{
Date of the day.
}
\description{
Get today French date in several formats.
}
\examples{
caractr::str_today_fr()
caractr::str_today_fr("litteral")

}
