% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordonnees.R
\name{str_validate_email}
\alias{str_validate_email}
\title{Validate email with a regex.}
\usage{
str_validate_email(email)
}
\arguments{
\item{email}{A email vector.}
}
\value{
A boolean vector.
}
\description{
Validate email with a regex.
}
\examples{
caractr::str_validate_email(c("test@test", "test@test.fr"))

}
