% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_validate_url}
\alias{str_validate_url}
\title{Validate a URL with a regex.}
\usage{
str_validate_url(url)
}
\arguments{
\item{url}{A URL vector.}
}
\value{
A boolean vector.
}
\description{
Validate a URL with a regex.
}
