% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordonnees.R
\name{valider_email_regex}
\alias{valider_email_regex}
\title{Valider des emails par regex}
\usage{
valider_email_regex(email)
}
\arguments{
\item{email}{Un vecteur d'emails}
}
\value{
Un vecteur de booléens.
}
\description{
Valider des emails par regex
}
\examples{
caractr::valider_email_regex("test@test")
caractr::valider_email_regex("test@test.fr")

}
