% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R
\name{conv_nombre_toutes_lettres}
\alias{conv_nombre_toutes_lettres}
\title{conv_nombre_toutes_lettres}
\usage{
conv_nombre_toutes_lettres(nombre, type = "lettre", langue = "fr")
}
\arguments{
\item{nombre}{\dots}

\item{type}{au choix : lettre|lettre_f|ieme|ieme_f|ieme_chiffre|ieme_chiffre_f}

\item{langue}{\dots}
}
\description{
conv_nombre_toutes_lettres
}
