% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{number_letter}
\alias{number_letter}
\title{A conversion table between numbers and their litteral forms in French.}
\format{A data frame with 1127 rows and 7 variables:
\describe{
  \item{x}{number, as integer}
  \item{letter}{number in its masculine litteral form}
  \item{letter_f}{number in its feminine litteral form}
  \item{ieme}{ordinal number in its masculine litteral form}
  \item{ieme_f}{ordinal number in its feminine litteral form}
  \item{ieme_number}{ordinal number in its masculine form}
  \item{ieme_number_f}{ordinal number in its feminine form}
}}
\usage{
number_letter
}
\description{
A conversion table between numbers and their litteral forms in French.
}
\keyword{datasets}
