% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{A list of stop words.}
\format{A data frame with 145 rows and 2 variables:
\describe{
  \item{language}{language code (ISO 639-1)}
  \item{stopword}{stop word}
}}
\usage{
stopwords
}
\description{
A list of stop words.
}
\keyword{datasets}
