% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_pretty_num}
\alias{str_pretty_num}
\title{Conversion from a numeric to a French formed number.}
\usage{
str_pretty_num(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector.
}
\description{
Includes comma as decimal mark and non-breaking space between thousands.
}
\examples{
caractr::str_pretty_num(1854.2)

}
