% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-word_fr_accent.R
\docType{data}
\name{word_fr_accent}
\alias{word_fr_accent}
\title{A conversion table between non accentued words and accentuated words in French.}
\format{A data frame with 13696 rows and 2 variables:
\describe{
  \item{word}{non accentued word}
  \item{word_fr_accent}{accentued word}
}}
\usage{
word_fr_accent
}
\description{
A conversion table between non accentued words and accentuated words in French.
}
\keyword{datasets}
